% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_efficient_frontier.R
\name{.aggregate_returns}
\alias{.aggregate_returns}
\title{Agrega retornos para frequência especificada}
\usage{
.aggregate_returns(ret_matrix, freq = "monthly", dates = NULL)
}
\description{
Agrega retornos para frequência especificada
}
\keyword{internal}
