% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brstocks-package.R
\docType{package}
\name{brstocks-package}
\alias{brstocks}
\alias{brstocks-package}
\title{brstocks: Brazilian Stock Market Data and Analysis}
\description{
Provides tools to retrieve and analyze historical stock data from the Brazilian stock exchange (B3) and other markets via Yahoo Finance. Includes functions for calculating returns, beta, correlation and covariance matrices, supporting quantitative finance analysis.
}
\author{
\strong{Maintainer}: Joao Paulo dos Santos Pereira Barbosa \email{joao.31582129@gmail.com}

Authors:
\itemize{
  \item Joao Paulo dos Santos Pereira Barbosa \email{joao.31582129@gmail.com}
}

}
\keyword{internal}
