% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_efficient_frontier.R
\name{.calc_expected_returns_from_matrix}
\alias{.calc_expected_returns_from_matrix}
\title{Calcula retornos esperados a partir da matriz}
\usage{
.calc_expected_returns_from_matrix(ret_matrix)
}
\description{
Calcula retornos esperados a partir da matriz
}
\keyword{internal}
